Split homepage into different component sections and delete config file
